humbugg/backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
click==8.1.7
python-jose==3.3.0
requests==2.31.0
pymongo==4.6.1
watchtower==3.0.1
boto3==1.28.72
python-dotenv==1.0.1
